Cocopen class created in cocopen.py file, run.py created
