+++ .gitignore
@@ -1,3 +1,5 @@
 .vscode/
 datasets/
 .env
+
+__pycache__
