+++ run.sh
@@ -1,0 +1,1 @@
+python ./src/coconut_v4.py
